backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
docker==7.1.0
requests==2.32.0
